% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tree.R
\name{prediction_df}
\alias{prediction_df}
\title{Apply the predicted tree on either new test data or training data.}
\usage{
prediction_df(fit, task, data)
}
\arguments{
\item{fit}{constparty object of the decision tree.}

\item{task}{Character. Task type: "classification" or "regression".}

\item{data}{A data.frame containing the features and target for prediction.}
}
\value{
A dataframe of prediction values with scaled columns
and clustered samples.
}
\description{
dplyr::select features with p-value (computed from decision tree) < \code{p_thres}
or all features if \code{show_all_feats == TRUE}.
}
\references{
\url{https://github.com/trangdata/treeheatr/blob/85be4a61e35a62285c95b553f03729721bb18a0b/R/utils.R}
}
