% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_norm}
\alias{scale_norm}
\title{Performs transformation on continuous variables.}
\usage{
scale_norm(x, trans_type = c("percentize", "normalize", "scale", "none"))
}
\arguments{
\item{x}{Numeric vector.}

\item{trans_type}{Character. One of "percentize","normalize","scale","none" passed to scale_norm().}
}
\value{
Numeric vector of the transformed \code{x}.
}
\description{
Performs transformation on continuous variables for the heatmap color scales.
}
\examples{
scale_norm(1:5, "normalize")

}
\references{
\url{https://github.com/trangdata/treeheatr/blob/85be4a61e35a62285c95b553f03729721bb18a0b/R/utils.R}
}
