% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted_mat.R
\name{supervised_distance_pair}
\alias{supervised_distance_pair}
\title{Calculate Supervised Distance Between Two Points}
\usage{
supervised_distance_pair(i, j, X, Y, linkage, representatives = NULL)
}
\arguments{
\item{i}{Index of first observation.}

\item{j}{Index of second observation.}

\item{X}{Numeric matrix of observations.}

\item{Y}{Vector of category labels.}

\item{linkage}{Character linkage method: "CT","SG","CP".}

\item{representatives}{list of R_ij and R_ji.}
}
\value{
Numeric distance value.
}
\description{
Computes the supervised distance metric combining within- and between-cluster distances.
}
