% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_all.R
\name{draw_all}
\alias{draw_all}
\title{Draw Full Visualization: Decision Tree with Heatmap and Evaluation}
\usage{
draw_all(
  prepare_tree,
  heat,
  total_w = 297,
  total_h = 210,
  layout,
  x_eval_start = 15,
  y_eval_start = NULL,
  eval_text = 7,
  eval_res = NULL,
  print_eval = TRUE,
  show_col_prox = TRUE,
  show_row_prox = TRUE
)
}
\arguments{
\item{prepare_tree}{A list returned from a tree preparation function,
containing \code{plot_data} and \code{branches} for tree structure.}

\item{heat}{A \code{grob} object representing the heatmap visualization. Usually generated with \code{grid.grabExpr(draw(...))}}

\item{total_w}{Total width of the drawing in mm. Default is 297 (A4 landscape width).}

\item{total_h}{Total height of the drawing in mm. Default is 210 (A4 landscape height).}

\item{layout}{A list specifying layout parameters: \code{tree_w}, \code{tree_h}, \code{margin}, and \code{offset_h}.}

\item{x_eval_start}{X-axis starting position (in mm) for evaluation text. Default is 15.}

\item{y_eval_start}{Y-axis starting position (in mm) for evaluation text. If NULL, it will be computed automatically.}

\item{eval_text}{Font size for the evaluation text. Default is 6.}

\item{eval_res}{A list with evaluation result text from \code{eval_tree()} (including \code{data_info}, \code{train_metrics}, and \code{test_metrics}).}

\item{print_eval}{Logical, whether to show evaluation results. Default is TRUE.}

\item{show_col_prox}{Logical, whether to show column proximity.}

\item{show_row_prox}{Logical, whether to show row proximity.}
}
\description{
This function creates a full-page layout consisting of a decision tree plot,
a heatmap, and optional evaluation results. It is designed for use in
reporting classification or clustering trees with additional visual indicators.
}
\examples{
library(rpart)
library(partykit)
library(ggparty)
library(dplyr)
library(seriation)
library(ComplexHeatmap)
library(circlize)
data_all <- add_data_type(data_train = train_covid, data_test = test_covid)
data <- prepare_features(data_all, target_lab = "Outcome", task = "classification")
train_tree = train_tree(data_train = train_covid, target_lab = "Outcome", model = "rpart")
fit = train_tree$fit
var_imp = train_tree$var_imp
tree_res = compute_tree(fit, model = "rpart", show ="test", data = data, target_lab = "Outcome", task = "classification")
sorted_dat = sorted_mat(tree_res, target_lab = "Outcome", show = "test")
split_vec = get_split_vec(sorted_dat, tree_res)
layout = compute_layout(sorted_dat)
row_prop_ha = row_prop_anno(sorted_dat, layout, split_vec)
pred_ha = prediction_annotation(sorted_dat,target_lab = "Outcome",
                                label_map = c("0" = "Survival", "1" = "Death"),
                                label_map_colors = c("Survival" = "#50046d", "Death" = "#fcc47f"))
eval_res = eval_tree(x = "covid", tree_res, target_lab = "Outcome", sorted_dat, show ="test", model = "rpart")
main_ht = make_main_heatmap(sorted_dat, split_vec, pred_ha, row_prop_ha, layout)
col_ht = col_ht(fit, sorted_dat, var_imp, layout)
combined_heatmap <- col_ht \%v\% main_ht
pals = my_get_palettes()
legends = generate_legend_bundle(sorted_dat = sorted_dat, type_cols = pals$type_cols, label_cols = pals$label_cols, prop_cols = pals$prop_cols,
           col_mat = pals$col_mat, col_Col_Proximity = pals$col_Col_Proximity, col_Row_Proximity = pals$col_Row_Proximity)
heat = grid.grabExpr(draw(combined_heatmap, heatmap_legend_list = legends, auto_adjust = FALSE, ht_gap = unit(0, "mm"), newpage = FALSE))
prepare_tree = prepare_tree(tree_res, model = "rpart")
draw_all(prepare_tree = prepare_tree, heat = heat, layout = layout, eval_res = eval_res)
}
