% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtGAP-package.R
\docType{package}
\name{dtGAP-package}
\alias{dtGAP-package}
\title{dtGAP: Supervised Generalized Association Plots Based on Decision Trees}
\description{
Enhances decision tree visualization by incorporating Generalized Association Plots (GAP) through matrix-based visualizations including confusion matrix maps, decision tree matrix maps, and predicted class membership maps based on supervised correlation and distance metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanmingwu1103/dtGAP}
  \item \url{https://CRAN.R-project.org/package=dtGAP}
  \item Report bugs at \url{https://github.com/hanmingwu1103/dtGAP/issues}
}

}
\author{
\strong{Maintainer}: Han-Ming Wu \email{wuhm@g.nccu.edu.tw}

Authors:
\itemize{
  \item Chia-Yu Chang \email{110304049@g.nccu.edu.tw}
  \item Chun-houh Chen \email{cchen@stat.sinica.edu.tw}
}

}
\keyword{internal}
