% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Psychosis_Disorder}
\alias{Psychosis_Disorder}
\title{Psychosis Disorder Data}
\format{
A data frame with 95 observations and 51 variables:
\describe{
\item{UNIQID}{Factor indicating disorder type.}
\item{AH1, AH2, AH3, AH4, AH5, AH6}{Hallucinations subscale (SAPS).}
\item{DL1, DL2, DL3, DL4, DL5, DL6, DL7, DL8, DL9, DL10, DL11, DL12}{Delusions subscale (SAPS).}
\item{BE1, BE2, BE3, BE4}{Behavior subscale (SAPS).}
\item{TH1, TH2, TH3, TH4, TH5, TH6, TH7, TH8}{Thought disorder subscale (SAPS).}
\item{NA1, NA2, NA3, NA4, NA5, NA6, NA7}{Expression subscale (SANS).}
\item{NB1, NB2, NB3, NB4}{Speech subscale (SANS).}
\item{NC1, NC2, NC3}{Hygiene subscale (SANS).}
\item{ND1, ND2, ND3, ND4}{Activity subscale (SANS).}
\item{NE1, NE2}{Inattentiveness subscale (SANS).}
}
}
\usage{
Psychosis_Disorder
}
\description{
Ratings of positive and negative symptoms in psychosis disorders,
based on Andreasen’s Scale for Assessment of Positive Symptoms (SAPS)
and Scale for Assessment of Negative Symptoms (SANS).
}
\details{
This data set comprises 95 subjects, of whom 69 were diagnosed with
schizophrenia and 26 with bipolar disorder. All symptoms were rated
on a six‐point scale (0–5).
}
\keyword{datasets}
